Create Exponential Search .c
Exponential search is also known as doubling or galloping search. This mechanism is used to find the range where the search key may present. If L and U are the upper and lower bound of the list, then L and U both are the power of 2. For the last section, the U is the last position of the list. For that reason, it is known as exponential.

After finding the specific range, it uses the binary search technique to find the exact location of the search key.

The complexity of Exponential Search Technique
Time Complexity: O(1) for the best case. O(log2 i) for average or worst case. Where i is the location where search key is present.
Space Complexity: O(1)